// src/app/interfaces/iposts.ts
export interface Iposts {
    post_id?: any;
    category_id?: any;
    account_id?: any;
    post_headline?: string;
    post_content?: string;
    post_time?: string;
    post_update?: string;
    post_type_id?: any;
    post_type_title?: string;
    post_views?: any;
    username?: string;
    message?: string;
}